feat(filters): implement FiltersBar and FiltersForm components for enhanced property search functionality
